refactoring - moving features to especialized files
